Implement frame table
This frame table adds 2 4GiB regions to the SOS address space. The first
holds the table of structures used to map the frames and the second is
where each of those frames are then mapped into SOS.

The frame table then maintains two doubly linked lists of free and
allocated frames.

As students will need to modify this code, there are numerous asserts to
ensure that the frame table is in a coherent state across all
operations.

Students will need to modify the frame table to implement some form of
clock replacement for demand paging.