Fix memory allocation for new untyped in bootinfo
Bootinfo now places all non-device memory at the end of the untyped list
and marks all physical memory addresses that are not physical memory or
in use by the kernel as device memory.

This change ensures that stolen memory is always physical memory and
that only the memory range mapped by the odroidc2 is added to the page
table.