cspace: Fix bugs in alloc and free slot cspace functions
The alloc slot function was ignoring one slot per bottom level cnode in
a 2-level cspace. That was >8000 total slots in the root task cspace!

The free slot function was computing the upper bound on CPtr values
incorrectly for 2-level cspaces.

Added a function to the cspace API for computing the current number of
capabilities stored in a cspace.