cspace: Fix bitfield bug introduced by 57d4b41
In the course of fixing other bugs, 57d4b41 moved the read of the bits
array outside the bounds checking of the index. The bug can manifest as
a failed assertion if the value on the stack preceding the bits array is
ULONG_MAX. In any case, it's probably best not to touch data outside the
bounds of the provided array.