Move thread registration from thread_trampoline to thread_create
Threads were registering themselves, which made us have to deal with
a more complicated concurrency model where a thread could call register
while the GDB thread was not in a position to handle it (as it cannot
control the execution of threads it is not aware of). This commit moves
thread registration into thread_spawn(), which should be called by a
controlled thread. As such, the debugger thread can prevent further
registrations until it is ready to deal with them, which simplifies its
implementation.

Signed-off-by: Alwin Joshy <joshyalwin@gmail.com>